KBTU/Lab4/requirements.txt
Django~=3.2.12
djangorestframework==3.11.2